justicebot/requirements.txt
# JusticeBot - Legal AI for Undertrial Prisoners
streamlit==1.37.1
streamlit-option-menu==0.3.14
streamlit-chat==0.2.0
langchain==0.2.1
langchain-community==0.2.1
langchain-groq==0.1.6
langchain-google-genai==1.0.1
chromadb==0.5.0
sentence-transformers==3.0.1
groq==0.9.0
google-generativeai==0.7.0
together==0.2.11
pandas==2.2.0
numpy==1.24.3
python-dotenv==1.0.0
plotly==5.22.0
matplotlib==3.8.4
seaborn==0.13.0
pypdf==4.0.1
fpdf2==2.7.1
pytesseract==0.3.10
pillow==10.2.0
opencv-python==4.9.0.80
pydub==0.25.1
pyaudio==0.2.13
requests==2.31.0
beautifulsoup4==4.12.3
lxml==4.9.4
click==8.1.7
tqdm==4.66.2
cachetools==5.3.2
joblib==1.4.0
pytest==7.4.3
black==24.1.1
flake8==7.0.0
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
flask-talisman==1.1.0
flask-limiter==3.5.0
flask-cors==4.0.0
bleach==6.1.0
python-dotenv==1.0.0
